refactor tests for window context adjustements
